Added .gitignore and updated VecEdit.py
