+++ VecEdit.py
@@ -1,12 +1,16 @@
 import sys
 import json
-from PySide6.QtWidgets import QApplication, QMainWindow, QFileDialog, QTreeView, QInputDialog, QPushButton, QGridLayout
-from PySide6.QtGui import QStandardItemModel, QStandardItem
-from PySide6.QtUiTools import QUiLoader
+from PySide6.QtWidgets import *
+from PySide6.QtGui import *
+from PySide6.QtUiTools import *
 import os
 import shutil
 import gzip
 import platform
+os.remove("./ve_log.log")
+def log_to_file(text):
+	with open("./ve_log.log", "a") as file:
+		file.write(f"{text}\n")
 
 json_data = {}
 light_stylesheet = """
@@ -18,7 +22,29 @@ light_stylesheet = """
 		background-color: lightgray;
 		color: black;
 	}
+	QTabWidget::pane {
+		border: 1px solid lightgray;
+	}
+	QTabBar::tab {
+		background: lightgray;
+		color: black;
+		padding: 5px;
+	}
+	QTabBar::tab:selected {
+		background: white;
+		border: 1px solid lightgray;
+		border-bottom-color: white;
+	}
+	QTreeView {
+		background-color: white;
+		color: black;
+	}
+	QTreeView::item:selected {
+		background-color: lightgray;
+		color: black;
+	}
 """
+
 
 dark_stylesheet = """
 	QWidget {
@@ -29,48 +55,85 @@ dark_stylesheet = """
 		background-color: #3d3d3d;
 		color: white;
 	}
+	QTabWidget::pane {
+		border: 1px solid #3d3d3d;
+	}
+	QTabBar::tab {
+		background: #3d3d3d;
+		color: white;
+		padding: 5px;
+	}
+	QTabBar::tab:selected {
+		background: #2d2d2d;
+		border: 1px solid #3d3d3d;
+		border-bottom-color: #2d2d2d;
+	}
+	QTreeView {
+		background-color: #2d2d2d;
+		color: white;
+	}
+	QTreeView::item:selected {
+		background-color: #3d3d3d;
+		color: white;
+	}
 """
-def detect_darkmode_in_windows(): 
+
+
+def detect_darkmode_in_windows():
+	log_to_file("Other function called.")
 	try:
 		import winreg
 	except ImportError:
 		return False
+	log_to_file("Winreg works")
 	registry = winreg.ConnectRegistry(None, winreg.HKEY_CURRENT_USER)
 	reg_keypath = r'SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize'
 	try:
 		reg_key = winreg.OpenKey(registry, reg_keypath)
 	except FileNotFoundError:
 		return False
+	log_to_file("Key imported")
 
 	for i in range(1024):
 		try:
 			value_name, value, _ = winreg.EnumValue(reg_key, i)
 			if value_name == 'AppsUseLightTheme':
+				log_to_file(f"Dark mode in Windows: {value == 0}")
 				return value == 0
 		except OSError:
 			break
+
+	log_to_file("Everything else broke.")
 	return False
 
 def detect_dark_mode():
+	log_to_file("Detecting dark mode")
 	if platform.system() == 'Linux':
 		try:
+			log_to_file("You are a Linux user.")
 			import subprocess
 			dark_mode = subprocess.check_output(
 				'gsettings get org.gnome.desktop.interface gtk-theme', shell=True).decode().strip()
+			log_to_file(f"Dark mode for Linux: {'dark' in dark_mode.lower()}")
 			return 'dark' in dark_mode.lower()
 		except:
 			pass
 	elif platform.system() == 'Windows':
+		log_to_file("You're a windows person. Moving detection to other function")
 		return detect_darkmode_in_windows()
 
 	elif platform.system() == 'Darwin':
+		log_to_file("ew macos")
 		try:
 			import subprocess
 			dark_mode = subprocess.check_output(
 				'ddefaults read -g AppleInterfaceStyle', shell=True).decode().strip()
+			log_to_file(f"Dark mode for $$$: {'dark' in dark_mode.lower()}")
 			return 'dark' in dark_mode.lower()
 		except:
 			pass
+
+	log_to_file("You're not using Windows, MacOS, or Linux. Why? WHY?")
 	return False
 
 
@@ -91,13 +154,9 @@ class MainWindow(QMainWindow):
 		ui_file_path = resource_path('main_window.ui')
 		self.ui = loader.load(ui_file_path, self)
 		
-		self.ui.import_button.clicked.connect(self.load_json_data)
-		self.ui.export_button.clicked.connect(self.export_json_data)
-		self.ui.JsonTree.doubleClicked.connect(self.edit_item)
-		
-		self.tree_model = QStandardItemModel()
-		self.ui.JsonTree.setModel(self.tree_model)
-		self.tree_model.setHorizontalHeaderLabels(['Key', 'Value'])
+		self.ui.ImportButton.clicked.connect(self.load_json_data)
+		self.ui.ExportButton.clicked.connect(self.export_json_data)
+
 
 	def load_json_data(self):
 		global json_data
@@ -118,7 +177,7 @@ class MainWindow(QMainWindow):
 			#vecedit_temp/example.gz from vec/saves/world_1.sav for example
 			shutil.copyfile(file_path, temp_gz_path)
 
-			#We want the JS to be the same name without the gz extension, so this strips the last 3 characters
+			#We want the JSON to be the same name without the gz extension, so this strips the last 3 characters
 			#vecedit_temp/example
 			temp_json_path = temp_gz_path[:-3]
 			print(f"Json path is {temp_json_path}")
@@ -127,127 +186,19 @@ class MainWindow(QMainWindow):
 			
 			with open(temp_json_path, 'r') as file:
 				json_data = json.load(file)
-				self.populate_tree_view(json_data)
 
-	def populate_tree_view(self, json_data, parent=None):
-		if parent is None:
-			self.tree_model.clear()
-			self.tree_model.setHorizontalHeaderLabels(['Key', 'Value'])
-			parent = self.tree_model.invisibleRootItem()
-
-		for key, value in json_data.items():
-			key_item = QStandardItem(key)
-			if isinstance(value, dict):
-				value_item = QStandardItem('[Object]')
-				parent.appendRow([key_item, value_item])
-				self.populate_tree_view(value, key_item)
-			elif isinstance(value, list):
-				value_item = QStandardItem('[Array]')
-				parent.appendRow([key_item, value_item])
-				self.populate_tree_view_list(value, key_item)
+			gamemode_data = json_data['GamemodeData']
+			gamemode_string = gamemode_data["ID"]
+			print(f"String is set. | {gamemode_string}")
+			
+			# Find the index of the gamemode_string
+			gamemode_index = self.ui.GamemodeInput.findText(gamemode_string)
+			if gamemode_index == -1:
+				print(f"Error: gamemode_string '{gamemode_string}' not found in GamemodeInput.")
 			else:
-				match value:
-					case str():
-						value_item = QStandardItem(f'[String] {value}')
+				self.ui.GamemodeInput.setCurrentIndex(gamemode_index)
 
-					case bool():
-						value_item = QStandardItem(f'[Boolean] {value}')
 
-					case int():
-						value_item = QStandardItem(f'[Integer] {value}')
-
-					case float():
-						value_item = QStandardItem(f'[Float] {value}')
-
-					case None:
-						value_item = QStandardItem('[NoneType] None')
-
-					case _:
-						print(type(value))
-						value_item = QStandardItem(str(value))
-
-				parent.appendRow([key_item, value_item])
-
-	def populate_tree_view_list(self, value_list, parent):
-		for i, value in enumerate(value_list):
-			index_item = QStandardItem(f'[{i}]')
-			if isinstance(value, dict):
-				value_item = QStandardItem('[Object]')
-				parent.appendRow([index_item, value_item])
-				self.populate_tree_view(value, index_item)
-			elif isinstance(value, list):
-				value_item = QStandardItem('[Array]')
-				parent.appendRow([index_item, value_item])
-				self.populate_tree_view_list(value, index_item)
-			else:
-				value_item = QStandardItem(str(value))
-				parent.appendRow([index_item, value_item])
-
-	def edit_item(self, index):
-		item = self.tree_model.itemFromIndex(index)
-		if item:
-			new_value, ok = QInputDialog.getText(self, "Edit Value", "Enter new value:", text=item.text())
-			if ok:
-				item.setText(new_value)
-				self.update_json_data()
-
-	def update_json_data(self, parent=None, json_obj=None):
-		if parent is None:
-			parent = self.tree_model.invisibleRootItem()
-			json_obj = json_data
-		for row in range(parent.rowCount()):
-			key_item = parent.child(row, 0)
-			value_item = parent.child(row, 1)
-			key = key_item.text()
-			value = value_item.text()
-			print(key, value)
-			if value == '[Object]':
-				json_obj[key] = {}	
-				self.update_json_data(key_item, json_obj[key])
-			elif value == '[Array]':
-				json_obj[key] = []
-				self.update_json_data_list(key_item, json_obj[key])
-
-			elif value.startswith('[Integer]'):
-				json_obj[key] = int(value.split(' ', 1)[1])
-
-			elif value.startswith('[Boolean]'):
-				json_obj[key] = value.split(' ', 1)[1] == 'True'
-
-			elif value.startswith('[Float]'):
-				json_obj[key] = float(value.split(' ', 1)[1])
-
-			elif value.startswith('[String]'):
-				json_obj[key] = str(value.split(' ', 1)[1])
-				
-			elif value == '[NoneType] None':
-				json_obj[key] = None
-
-			else:
-				json_obj[key] = value
-
-	def update_json_data_list(self, parent, json_list):
-		for row in range(parent.rowCount()):
-			index_item = parent.child(row, 0)
-			value_item = parent.child(row, 1)
-			index = int(index_item.text().strip('[]'))
-			value = value_item.text()
-			if value == '[Object]':
-				json_list.append({})
-				self.update_json_data(index_item, json_list[index])
-			elif value == '[Array]':
-				json_list.append([])
-				self.update_json_data_list(index_item, json_list[index])
-			elif value.startswith('[Integer]'):
-				json_list.append(int(value.split(' ', 1)[1]))
-			elif value.startswith('[Boolean]'):
-				json_list.append(value.split(' ', 1)[1] == 'True')
-			elif value.startswith('[Float]'):
-				json_list.append(float(value.split(' ', 1)[1]))
-			elif value == '[NoneType] None':
-				json_list.append(None)
-			else:
-				json_list.append(value)
 
 	def export_json_data(self):
 		global json_data
